Move TreeBPE core components to inner module
